+++ _examples/helloworld/producer/main.go
@@ -8,11 +8,12 @@ import (
 	"os"
 	"time"
 
-	"github.com/Azure/azure-amqp-common-go/aad"
-	"github.com/Azure/azure-event-hubs-go"
+	"github.com/Azure/azure-amqp-common-go/v2/aad"
 	mgmt "github.com/Azure/azure-sdk-for-go/services/eventhub/mgmt/2017-04-01/eventhub"
 	"github.com/Azure/go-autorest/autorest/azure"
 	azauth "github.com/Azure/go-autorest/autorest/azure/auth"
+
+	"github.com/Azure/azure-event-hubs-go/v2"
 )
 
 const (