+++ amqp_mgmt.go
@@ -27,9 +27,9 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/Azure/azure-amqp-common-go/rpc"
+	"github.com/Azure/azure-amqp-common-go/v2/rpc"
+	"github.com/devigned/tab"
 	"github.com/mitchellh/mapstructure"
-	"go.opencensus.io/trace"
 	"pack.ag/amqp"
 )
 
@@ -83,7 +83,7 @@ func newClient(namespace *namespace, hubName string) *client {
 
 // GetHubRuntimeInformation requests runtime information for an Event Hub
 func (c *client) GetHubRuntimeInformation(ctx context.Context, conn *amqp.Client) (*HubRuntimeInformation, error) {
-	ctx, span := trace.StartSpan(ctx, "eh.mgmt.client.GetHubRuntimeInformation")
+	ctx, span := tab.StartSpan(ctx, "eh.mgmt.client.GetHubRuntimeInformation")
 	defer span.End()
 
 	rpcLink, err := rpc.NewLink(conn, address)
@@ -117,7 +117,7 @@ func (c *client) GetHubRuntimeInformation(ctx context.Context, conn *amqp.Client
 
 // GetHubPartitionRuntimeInformation fetches runtime information from the AMQP management node for a given partition
 func (c *client) GetHubPartitionRuntimeInformation(ctx context.Context, conn *amqp.Client, partitionID string) (*HubPartitionRuntimeInformation, error) {
-	ctx, span := trace.StartSpan(ctx, "eh.mgmt.client.GetHubPartitionRuntimeInformation")
+	ctx, span := tab.StartSpan(ctx, "eh.mgmt.client.GetHubPartitionRuntimeInformation")
 	defer span.End()
 
 	rpcLink, err := rpc.NewLink(conn, address)