+++ hub.go
@@ -33,19 +33,20 @@ import (
 	"path"
 	"sync"
 
-	"github.com/Azure/azure-amqp-common-go/aad"
-	"github.com/Azure/azure-amqp-common-go/auth"
-	"github.com/Azure/azure-amqp-common-go/conn"
-	"github.com/Azure/azure-amqp-common-go/log"
-	"github.com/Azure/azure-amqp-common-go/persist"
-	"github.com/Azure/azure-amqp-common-go/sas"
+	"github.com/Azure/azure-amqp-common-go/v2/aad"
+	"github.com/Azure/azure-amqp-common-go/v2/auth"
+	"github.com/Azure/azure-amqp-common-go/v2/conn"
+	"github.com/Azure/azure-amqp-common-go/v2/sas"
+	"github.com/Azure/azure-amqp-common-go/v2/uuid"
 	"github.com/Azure/azure-sdk-for-go/services/eventhub/mgmt/2017-04-01/eventhub"
 	"github.com/Azure/go-autorest/autorest/azure"
 	"github.com/Azure/go-autorest/autorest/date"
 	"github.com/Azure/go-autorest/autorest/to"
+	"github.com/devigned/tab"
 	"pack.ag/amqp"
 
-	"github.com/Azure/azure-event-hubs-go/atom"
+	"github.com/Azure/azure-event-hubs-go/v2/atom"
+	"github.com/Azure/azure-event-hubs-go/v2/persist"
 )
 
 const (
@@ -216,7 +217,7 @@ func (hm *HubManager) Put(ctx context.Context, name string, opts ...HubManagemen
 
 	reqBytes, err := xml.Marshal(he)
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return nil, err
 	}
 
@@ -227,13 +228,13 @@ func (hm *HubManager) Put(ctx context.Context, name string, opts ...HubManagemen
 	}
 
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return nil, err
 	}
 
 	b, err := ioutil.ReadAll(res.Body)
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return nil, err
 	}
 
@@ -256,13 +257,13 @@ func (hm *HubManager) List(ctx context.Context) ([]*HubEntity, error) {
 	}
 
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return nil, err
 	}
 
 	b, err := ioutil.ReadAll(res.Body)
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return nil, err
 	}
 
@@ -290,7 +291,7 @@ func (hm *HubManager) Get(ctx context.Context, name string) (*HubEntity, error) 
 	}
 
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return nil, err
 	}
 
@@ -300,7 +301,7 @@ func (hm *HubManager) Get(ctx context.Context, name string) (*HubEntity, error) 
 
 	b, err := ioutil.ReadAll(res.Body)
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return nil, err
 	}
 
@@ -489,19 +490,19 @@ func (h *Hub) GetRuntimeInformation(ctx context.Context) (*HubRuntimeInformation
 	client := newClient(h.namespace, h.name)
 	c, err := h.namespace.newConnection()
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return nil, err
 	}
 
 	defer func() {
 		if err := c.Close(); err != nil {
-			log.For(ctx).Error(err)
+			tab.For(ctx).Error(err)
 		}
 	}()
 
 	info, err := client.GetHubRuntimeInformation(ctx, c)
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return nil, err
 	}
 
@@ -515,13 +516,13 @@ func (h *Hub) GetPartitionInformation(ctx context.Context, partitionID string) (
 	client := newClient(h.namespace, h.name)
 	c, err := h.namespace.newConnection()
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return nil, err
 	}
 
 	defer func() {
 		if err := c.Close(); err != nil {
-			log.For(ctx).Error(err)
+			tab.For(ctx).Error(err)
 		}
 	}()
 
@@ -542,12 +543,12 @@ func (h *Hub) Close(ctx context.Context) error {
 		if err := h.sender.Close(ctx); err != nil {
 			if rErr := h.closeReceivers(ctx); rErr != nil {
 				if !isConnectionClosed(rErr) {
-					log.For(ctx).Error(rErr)
+					tab.For(ctx).Error(rErr)
 				}
 			}
 
 			if !isConnectionClosed(err) {
-				log.For(ctx).Error(err)
+				tab.For(ctx).Error(err)
 				return err
 			}
 
@@ -558,7 +559,7 @@ func (h *Hub) Close(ctx context.Context) error {
 	// close receivers and return error
 	err := h.closeReceivers(ctx)
 	if err != nil && !isConnectionClosed(err) {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return err
 	}
 
@@ -573,7 +574,7 @@ func (h *Hub) closeReceivers(ctx context.Context) error {
 	var lastErr error
 	for _, r := range h.receivers {
 		if err := r.Close(ctx); err != nil {
-			log.For(ctx).Error(err)
+			tab.For(ctx).Error(err)
 			lastErr = err
 		}
 	}
@@ -610,7 +611,7 @@ func (h *Hub) Receive(ctx context.Context, partitionID string, handler Handler, 
 	// Todo: change this to use name rather than identifier
 	if r, ok := h.receivers[receiver.getIdentifier()]; ok {
 		if err := r.Close(ctx); err != nil {
-			log.For(ctx).Error(err)
+			tab.For(ctx).Error(err)
 		}
 	}
 
@@ -635,24 +636,49 @@ func (h *Hub) Send(ctx context.Context, event *Event, opts ...SendOption) error 
 	return sender.Send(ctx, event, opts...)
 }
 
-// SendBatch sends an EventBatch to the Event Hub
-//
-// SendBatch will retry sending the message for as long as the context allows
-func (h *Hub) SendBatch(ctx context.Context, batch *EventBatch, opts ...SendOption) error {
+// SendBatch sends a batch of events to the Hub
+func (h *Hub) SendBatch(ctx context.Context, iterator BatchIterator, opts ...BatchOption) error {
 	span, ctx := h.startSpanFromContext(ctx, "eh.Hub.SendBatch")
 	defer span.End()
 
 	sender, err := h.getSender(ctx)
 	if err != nil {
+		tab.For(ctx).Error(err)
 		return err
 	}
 
-	event, err := batch.toEvent()
-	if err != nil {
-		return err
+	batchOptions := &BatchOptions{
+		MaxSize: DefaultMaxMessageSizeInBytes,
 	}
 
-	return sender.Send(ctx, event, opts...)
+	for _, opt := range opts {
+		if err := opt(batchOptions); err != nil {
+			tab.For(ctx).Error(err)
+			return err
+		}
+	}
+
+	for !iterator.Done() {
+		id, err := uuid.NewV4()
+		if err != nil {
+			tab.For(ctx).Error(err)
+			return err
+		}
+
+		batch, err := iterator.Next(id.String(), batchOptions)
+
+		if err != nil {
+			tab.For(ctx).Error(err)
+			return err
+		}
+
+		if err := sender.trySend(ctx, batch); err != nil {
+			tab.For(ctx).Error(err)
+			return err
+		}
+	}
+
+	return nil
 }
 
 // HubWithPartitionedSender configures the Hub instance to send to a specific event Hub partition
@@ -720,7 +746,7 @@ func (h *Hub) getSender(ctx context.Context) (*sender, error) {
 	if h.sender == nil {
 		s, err := h.newSender(ctx)
 		if err != nil {
-			log.For(ctx).Error(err)
+			tab.For(ctx).Error(err)
 			return nil, err
 		}
 		h.sender = s