+++ storage/credential_test.go
@@ -29,9 +29,10 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/Azure/azure-event-hubs-go/internal/test"
 	"github.com/Azure/azure-storage-blob-go/azblob"
 	"github.com/stretchr/testify/suite"
+
+	"github.com/Azure/azure-event-hubs-go/v2/internal/test"
 )
 
 type (