+++ changelog.md
@@ -1,5 +1,13 @@
 # Change Log
 
+
+## `v2.0.0`
+- **breaking change:** moved github.com/Azure/azure-amqp-common-go/persist to
+  github.com/Azure/azure-event-hubs-go/persist
+- **breaking change:** changed batch message sending to use a safe batch iterator rather than leaving batch sizing to
+  the consumer.
+- move tracing to devigned/tab so to not have to take a direct dependency on opentracing or opencensus
+
 ## `v1.3.1`
 - cleanup connection after making management request